Move AddTorrentParams to its own header
